Architectural refactoring of async execute API
